Updates deps and fixes pre commits
