--- .markdownlint.json
@@ -1,15 +1,0 @@
-{
-    "default": true,
-    "MD004": {
-        "style": "dash"
-    },
-    "MD013": {
-        "line_length": 1e6
-    },
-    "MD029": {
-        "style": "ordered"
-    },
-    "MD033": false,
-    "MD041": false,
-    "MD002": false
-}
