+++ .prettierignore
@@ -1,0 +1,6 @@
+*.hbs
+.eslintignore
+.editorconfig
+.gitignore
+yarn.lock
+.prettierignore
