+++ generators/logger.ts
@@ -1,0 +1,3 @@
+import Logger from '@wildebeest/Logger';
+
+export default new Logger();
