+++ generators/migrationTypes/tableColumnsMigration/lib.ts
@@ -1,11 +1,27 @@
-// migrationTypes
-import { SEQUELIZE_ENUM_NAME_REGEX } from '../changeEnumAttributes/regexs';
+/**
+ * Extract the name and type of enum from an attribute definition
+ */
+export const SEQUELIZE_ENUM_NAME_REGEX = /DataTypes\.ENUM,\n {2}values: Object\.values\((.+?)\)/;
 
-// local
-const {
-  ATTRIBUTE_VALUE_REGEX,
-  ATTRIBUTE_VALIDATOR_REGEX,
-} = require('./regexs');
+/**
+ * Detect if an attribute definition has a validator
+ */
+export const ATTRIBUTE_VALIDATOR_REGEX = /validate: ([A-Z_].+?),/;
+
+/**
+ * Create an attribute regex
+ *
+ * @param name - The name of the attribute
+ * @param exportType - The manner in which the attribute is exported (i.e. export const )
+ * @returns The regex for matching comment and definition from an attribute
+ */
+export const ATTRIBUTE_VALUE_REGEX = (
+  name: string,
+  exportType: string,
+): RegExp =>
+  new RegExp(
+    `\\/\\*\\*\\n \\* (.+?)\\n \\*[\\s\\S]+${exportType}${name}: ModelAttributeColumnOptions = {([\\s\\S]+?)};`,
+  );
 
 /**
  * Replace part of the definition using a regex
