--- generators/migrationTypes/tablesMigration/regexs.ts
@@ -1,4 +1,0 @@
-/**
- * The name of a class
- */
-export const CLASS_NAME_REGEX = /@class (.*)/;
