+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@transcend-io/wildebeest",
-  "version": "2.14.1",
+  "version": "2.14.2",
   "description": "Type-safe sequelize with a simplified migration framework",
   "main": "build/src/index.js",
   "homepage": "https://github.com/transcend-io/wildebeest#readme",
@@ -39,7 +39,7 @@
     "build:watch": "tsc --watch"
   },
   "dependencies": {
-    "aws-sdk": "^2.630.0",
+    "aws-sdk": "^2.639.0",
     "change-case": "=3.1.0",
     "express": "^4.17.1",
     "handlebars": "^4.7.3",
@@ -49,20 +49,20 @@
     "umzug": "^2.2.0"
   },
   "devDependencies": {
-    "@types/express": "^4.17.2",
+    "@types/express": "^4.17.3",
     "@types/lodash": "^4.14.149",
     "@types/mkdirp": "^1.0.0",
-    "@types/mocha": "^7.0.1",
-    "@types/node": "^13.7.7",
+    "@types/mocha": "^7.0.2",
+    "@types/node": "^13.9.1",
     "@types/pluralize": "0.0.29",
     "@types/umzug": "^2.2.3",
     "@types/uuid": "^7.0.0",
-    "@typescript-eslint/eslint-plugin": "^2.21.0",
-    "@typescript-eslint/parser": "^2.21.0",
+    "@typescript-eslint/eslint-plugin": "^2.23.0",
+    "@typescript-eslint/parser": "^2.23.0",
     "babel-eslint": "^10.1.0",
     "copy": "^0.3.2",
     "eslint": "^6.8.0",
-    "eslint-config-airbnb-base": "^14.0.0",
+    "eslint-config-airbnb-base": "^14.1.0",
     "eslint-config-prettier": "^6.10.0",
     "eslint-import-resolver-typescript": "^2.0.0",
     "eslint-plugin-import": "^2.20.1",
@@ -70,7 +70,7 @@
     "mocha": "^7.1.0",
     "prettier": "^1.19.1",
     "rimraf": "^3.0.2",
-    "tslint": "^6.0.0",
+    "tslint": "^6.1.0",
     "tslint-microsoft-contrib": "^6.2.0",
     "typescript": "^3.8.3"
   }
