+++ tslint.json
@@ -6,7 +6,8 @@
   "rules": {
     "array-type": false,
     "camel-case": false,
-    "completed-docs": [true,
+    "completed-docs": [
+      true,
       "classes",
       "enums",
       "enum-members",
@@ -17,28 +18,28 @@
       "properties",
       "types"
     ],
-      "export-name": false,
-      "interface-over-type-literal": false,
-      "max-line-length": false,
-      "new-parens": true,
-      "prefer-for-of": false,
-      "no-arg": true,
-      "ordered-imports": true,
-      "object-literal-key-quotes": false,
-      "no-bitwise": true,
-      "no-conditional-assignment": true,
-      "no-consecutive-blank-lines": false,
-      "no-relative-imports": [true, "allow-siblings"],
-      "no-console": false,
-      "no-unused-expression": false,
-      "no-namespace": false,
-      "no-shadowed-variable": false,
-      "no-var-requires": false,
-      "object-literal-sort-keys":false,
-      "quotemark": [false, "single", "avoid-escape", "avoid-template"],
-      "variable-name": false,
-      "trailing-comma": false,
-      "semicolon": false
+    "export-name": false,
+    "interface-over-type-literal": false,
+    "max-line-length": false,
+    "new-parens": true,
+    "prefer-for-of": false,
+    "no-arg": true,
+    "ordered-imports": true,
+    "object-literal-key-quotes": false,
+    "no-bitwise": true,
+    "no-conditional-assignment": true,
+    "no-consecutive-blank-lines": false,
+    "no-relative-imports": [true, "allow-siblings"],
+    "no-console": false,
+    "no-unused-expression": false,
+    "no-namespace": false,
+    "no-shadowed-variable": false,
+    "no-var-requires": false,
+    "object-literal-sort-keys": false,
+    "quotemark": [false, "single", "avoid-escape", "avoid-template"],
+    "variable-name": false,
+    "trailing-comma": false,
+    "semicolon": false
   },
   "jsRules": {
     "array-type": false,
@@ -56,6 +57,11 @@
     "semicolon": false
   },
   "linterOptions": {
-    "exclude": ["node_modules/**/*.ts", "dist/**/*.ts", "build/**/*.js", "builds/**/*.js"]
+    "exclude": [
+      "node_modules/**/*.ts",
+      "dist/**/*.ts",
+      "build/**/*.js",
+      "builds/**/*.js"
+    ]
   }
 }
